Potentially make add_tile faster by adding then removing
Removing first could take the reference count to 0, and add it to the GC
queue to be processed at GC time. However, if we're then adding the same
tile back on, it doesn't get removed from the GC queue.

By adding the new tile on, and then removing the old one, if it's the
same tile, we just increase the reference count to 2 and then back down
to 1.

Setting a tile to itself is a reasonably common operation, and it isn't
that uncommon for tiles to be unique across the screen too, so hopefully
this'll improve performance and either way is probably the cause of the
incredibly bad performance of the old tile rendering before the more
recent refactor (in addition to the other horribly inefficient code that
was there).