Do tile GC at a more sensible time
We were running tile GC before updating the actual tile data, because 
the `commit_data` may hold the last reference to a given tile. But if we
don't `.take()` it and instead `.as_ref()`, then while the `.gc()` call 
is being done, the references are still alive.

So we can now run gc _after_ doing the tile update which is much better 
because the important bit is that the tiles are updated first.